Add ammo & reloading

+++ server/src/objects/loot.ts
@@ -9,13 +9,17 @@ import { v2v } from "../utils/misc";
 
 import { type SuroiBitStream } from "../../../common/src/utils/suroiBitStream";
 import { type ObjectType } from "../../../common/src/utils/objectType";
-import { type Vector } from "../../../common/src/utils/vector";
-import { randomBoolean, randomRotation } from "../../../common/src/utils/random";
+import {
+    v, vAdd, type Vector
+} from "../../../common/src/utils/vector";
+import { randomRotation } from "../../../common/src/utils/random";
 import { type LootDefinition } from "../../../common/src/definitions/loots";
 import { ItemType } from "../../../common/src/utils/objectDefinitions";
 import { type Player } from "./player";
 import { CircleHitbox } from "../../../common/src/utils/hitbox";
 import { HealType } from "../../../common/src/definitions/healingItems";
+import { PickupPacket } from "../packets/sending/pickupPacket";
+import { MaxInventoryCapacity } from "../../../common/src/constants";
 
 export class Loot extends GameObject {
     override readonly is: CollisionFilter = {
@@ -35,23 +39,44 @@ export class Loot extends GameObject {
     body: Body;
 
     oldPosition: Vector;
-    oldRotation = 0;
+
+    count = 1;
 
     isNew = true;
 
-    constructor(game: Game, type: ObjectType, position: Vector) {
+    constructor(game: Game, type: ObjectType, position: Vector, count?: number) {
         super(game, type, position);
 
         this.oldPosition = position;
+        if (count !== undefined) this.count = count;
 
-        // Create the body
+        // Create the body and hitbox
         this.body = game.world.createBody({
             type: "dynamic",
             position: v2v(position),
             linearDamping: 0.003,
-            angularDamping: 0
+            angularDamping: 0,
+            fixedRotation: true
         });
-        const radius = (this.type.definition as LootDefinition).itemType === ItemType.Gun ? 3.125 : 2.5;
+        const itemType = (this.type.definition as LootDefinition).itemType;
+        let radius: number;
+        switch (itemType) {
+            case ItemType.Gun:
+                radius = 3.4;
+                break;
+            case ItemType.Ammo:
+                radius = 2;
+                break;
+            case ItemType.Melee:
+                radius = 3;
+                break;
+            case ItemType.Healing:
+                radius = 2.5;
+                break;
+            default:
+                radius = 2.5;
+                break;
+        }
         this.body.createFixture({
             shape: Circle(radius),
             restitution: 0,
@@ -59,13 +84,11 @@ export class Loot extends GameObject {
             friction: 0.0,
             userData: this
         });
-
         this.hitbox = new CircleHitbox(radius, this.position);
 
         // Push the loot in a random direction
         const angle = randomRotation();
         this.body.setLinearVelocity(Vec2(Math.cos(angle), Math.sin(angle)).mul(0.005));
-        this.body.applyTorque(randomBoolean() ? 0.003 : -0.003);
 
         setTimeout((): void => { this.isNew = false; }, 100);
     }
@@ -80,6 +103,7 @@ export class Loot extends GameObject {
     }
 
     canInteract(player: Player): boolean {
+        if (this.dead) return false;
         const inventory = player.inventory;
         const definition = this.type.definition as LootDefinition;
         switch (definition.itemType) {
@@ -96,53 +120,80 @@ export class Loot extends GameObject {
                     !inventory.hasWeapon(1) ||
                     (inventory.activeWeaponIndex < 2 && this.type.idNumber !== inventory.activeWeapon.type.idNumber);
             }
+            case ItemType.Ammo: {
+                const idString = this.type.idString;
+                const currentCount: number = inventory.items[idString];
+                const maxCapacity: number = MaxInventoryCapacity[idString];
+                return currentCount + 1 <= maxCapacity;
+            }
             case ItemType.Melee: {
                 return this.type.idNumber !== inventory.getWeapon(2)?.type.idNumber;
             }
         }
+        return false;
     }
 
     interact(player: Player): void {
         const inventory = player.inventory;
-        let success = false;
-
+        let deleteItem = true;
         const definition = this.type.definition as LootDefinition;
 
         switch (definition.itemType) {
             case ItemType.Healing: {
-                success = true;
-
                 if (definition.healType === HealType.Health) player.health += definition.restoreAmount;
                 else if (definition.healType === HealType.Adrenaline) player.adrenaline += definition.restoreAmount;
-
                 break;
             }
             case ItemType.Melee: {
                 inventory.addOrReplaceWeapon(2, this.type.idString);
-                success = true;
                 break;
             }
             case ItemType.Gun: {
                 if (!inventory.hasWeapon(0) || !inventory.hasWeapon(1)) {
                     inventory.appendWeapon(this.type.idString);
-                    success = true;
                 } else if (inventory.activeWeaponIndex < 2 && this.type.idString !== inventory.activeWeapon.type.idString) {
                     inventory.addOrReplaceWeapon(inventory.activeWeaponIndex, this.type.idString);
-                    success = true;
                 }
                 break;
             }
+            case ItemType.Ammo: {
+                const idString = this.type.idString;
+                const currentCount: number = inventory.items[idString];
+                const maxCapacity: number = MaxInventoryCapacity[idString];
+
+                if (currentCount + this.count <= maxCapacity) {
+                    inventory.items[idString] += this.count;
+                } else if (currentCount + 1 > maxCapacity) {
+                    // inventory full
+                } else if (currentCount + this.count > maxCapacity) {
+                    inventory.items[idString] = maxCapacity;
+                    this.count = (currentCount + this.count) - maxCapacity;
+                    this.game.fullDirtyObjects.add(this);
+                    deleteItem = false;
+                }
+                console.log(inventory.items);
+                break;
+            }
         }
-        if (success) {
-            this.game.dynamicObjects.delete(this);
-            this.game.loot.delete(this);
-            this.game.removeObject(this);
-            this.game.world.destroyBody(this.body);
-        }/* else if (!ignoreItem) {
+
+        // Destroy the old loot
+        this.game.dynamicObjects.delete(this);
+        this.game.loot.delete(this);
+        this.game.removeObject(this);
+        this.game.world.destroyBody(this.body);
+        player.dirty.inventory = true;
+        this.dead = true;
+
+        // Send pickup packet
+        if (definition.itemType !== ItemType.Gun) {
+            player.sendPacket(new PickupPacket(player));
+        }
+
+        // If the item wasn't deleted, create a new loot item pushed slightly away from the player
+        if (!deleteItem) {
             const invertedAngle = (player.rotation + Math.PI) % (2 * Math.PI);
-            /* eslint-disable-next-line no-new
-            new Loot(this.game, this.type, vAdd(this.position, v(0.4 * Math.cos(invertedAngle), 0.4 * Math.sin(invertedAngle))));
-        } */
+            this.game.addLoot(this.type, vAdd(this.position, v(0.4 * Math.cos(invertedAngle), 0.4 * Math.sin(invertedAngle))), this.count);
+        }
     }
 
     // eslint-disable-next-line @typescript-eslint/no-empty-function
@@ -150,7 +201,6 @@ export class Loot extends GameObject {
 
     override serializePartial(stream: SuroiBitStream): void {
         stream.writePosition(this.position);
-        stream.writeRotation(this.rotation, 8);
     }
 
     override serializeFull(stream: SuroiBitStream): void {
